AbstractBean
