+++ mysql/src/main/java/dbdemo/mysql/entity/User.java
@@ -1,23 +1,15 @@
 package dbdemo.mysql.entity;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
-import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
-import java.util.Date;
 import java.util.List;
 
 @Entity
 @Table(name = "user")
-public class User implements java.io.Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+public class User extends AbstractBean implements java.io.Serializable {
 
     private String name;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private Date createDate;
 
     @ManyToOne
     @JoinColumn(name = "did")
@@ -30,31 +22,12 @@ public class User implements java.io.Serializable {
             inverseJoinColumns = {@JoinColumn(name = "roles_id")})
     private List<Role> roles;
 
-    public User() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
     public String getName() {
         return name;
     }
 
     public void setName(String name) {
         this.name = name;
-    }
-
-    public Date getCreateDate() {
-        return createDate;
-    }
-
-    public void setCreateDate(Date createDate) {
-        this.createDate = createDate;
     }
 
     public Department getDepartment() {
